refactoring
create a new folder for commands inside command
