+++ src/main/java/seedu/typists/command/commands/HistoryCommand.java
@@ -1,5 +1,6 @@
-package seedu.typists.command;
+package seedu.typists.command.commands;
 
+import seedu.typists.command.commands.Command;
 import seedu.typists.common.exception.InvalidCommandException;
 import seedu.typists.game.GameRecord;
 import seedu.typists.game.GameRecordsManager;
